# Faster utf7 encode