~40% faster for input with a mix of unicode and ascii chars. The improvement is more significant for pure ascii chars.

This is achieved by removing the .encode() in the most common path (ascii char) and comparing ordinal values instead of unicode objects.

Actually we can get 10% more by eliminating dot notation, 10-20% more by inlining `consume_b64_buffer` function, some more by caching `ord` global function, but then the code will get quite ugly and I'm not sure if that's welcomed.