@carsonip 2019-03-25
As a side note, combining with Cython, this can achieve a total of 10x improvement.