@carsonip 2019-04-05
Forgot to attach the benchmark script and results:
```
# -*- coding: utf-8 -*-
import time
from imapclient.imap_utf7 import encode, decode

q = u'你好嗎你好嗎ab 123&123 abc你' * 50

start = time.time()
for i in xrange(5000):
   encode(q)
print(time.time() - start)
```

CPython2.7
before: 2.45s
after: 1.48s

PyPy2.7 v7.1
before: 0.284s
after: 0.198s